Fix scroll reset bug in NepoNet; end NepoNet at bottom; add ideas; more convention center events; add gold globe
The scroll would reset to 0 if you keep clicking while it's already scrolling, as long as the top row stays completed. No longer!
I made the NepoNet grid long enough that you shouldn't be able to reach the end easily, but I still do it fairly often, so instead of fighting the player, I decided to just make it end on its own once all visible tiles are connected.