Fix some water treatment display *and* actual calculations; slightly reduce water treatment amount
The slight reduced amount is partly so the short ticks use a whole number and partly so the player sees a lower amount of water treatment than they actually get, kind of acting as a buffer. Also because it was like 6x more water treatment than my comment in the code said. :)